refactor(cli): Auto-validate after adding a source
