+++ crates/coral-cli/src/onboard.rs
@@ -1,6 +1,4 @@
-use std::collections::BTreeMap;
-
-use coral_api::v1::{AvailableSource, ExecuteSqlRequest, Source, ValidateSourceResponse};
+use coral_api::v1::{AvailableSource, ExecuteSqlRequest, Source};
 use coral_client::{
     AppClient, decode_execute_sql_response, default_workspace, format_batches_table,
 };
@@ -156,7 +154,12 @@ async fn run_installed_source_menu(
 
     match selection.map(|i| actions[i]) {
         Some(InstalledSourceAction::Validate) => {
-            validate_after_install(app, &source.name).await?;
+            source_ops::validate_and_print(
+                app,
+                &source.name,
+                Some(source_ops::MAX_TABLES_PER_SCHEMA),
+            )
+            .await?;
         }
         Some(InstalledSourceAction::Reconfigure) => {
             let inputs = source
@@ -168,7 +171,12 @@ async fn run_installed_source_menu(
             let result =
                 source_ops::add_bundled_source(app, &source.name, variables, secrets).await?;
             println!("Reconfigured source {}", result.name);
-            validate_after_install(app, &result.name).await?;
+            source_ops::validate_and_print(
+                app,
+                &result.name,
+                Some(source_ops::MAX_TABLES_PER_SCHEMA),
+            )
+            .await?;
         }
         Some(InstalledSourceAction::Back) | None => {}
     }
@@ -188,73 +196,7 @@ async fn run_add_bundled_source(
     let (variables, secrets) = source_ops::prompt_for_inputs(&inputs)?;
     let result = source_ops::add_bundled_source(app, &source.name, variables, secrets).await?;
     println!("Added source {}", result.name);
-    validate_after_install(app, &result.name).await
-}
-
-async fn validate_after_install(app: &AppClient, source_name: &str) -> Result<(), anyhow::Error> {
-    let response = source_ops::validate_source(app, source_name).await?;
-    print_validation_pretty(&response)
-}
-
-const MAX_TABLES_PER_SCHEMA: usize = 9;
-
-fn print_validation_pretty(response: &ValidateSourceResponse) -> Result<(), anyhow::Error> {
-    let source = response
-        .source
-        .as_ref()
-        .ok_or_else(|| anyhow::anyhow!("validate response missing source metadata"))?;
-
-    println!();
-    println!(
-        "  {} {}",
-        style("✓").green(),
-        style(format!("{} connected successfully", source.name)).bold()
-    );
-
-    // Group tables by schema, sorted.
-    let mut by_schema: BTreeMap<&str, Vec<&str>> = BTreeMap::new();
-    for table in &response.tables {
-        by_schema
-            .entry(&table.schema_name)
-            .or_default()
-            .push(&table.name);
-    }
-    for tables in by_schema.values_mut() {
-        tables.sort_unstable();
-    }
-
-    for (schema, tables) in &by_schema {
-        let count = tables.len();
-        println!();
-        println!(
-            "    {}",
-            style(format!(
-                "{schema} ({count} {})",
-                if count == 1 { "table" } else { "tables" }
-            ))
-            .bold()
-        );
-
-        let show_count = tables.len().min(MAX_TABLES_PER_SCHEMA);
-        let remaining = tables.len() - show_count;
-
-        for (i, table) in tables.iter().take(show_count).enumerate() {
-            let is_last = i == show_count - 1 && remaining == 0;
-            let branch = if is_last { "└─" } else { "├─" };
-            println!("    {} {}", style(branch).dim(), table);
-        }
-
-        if remaining > 0 {
-            println!(
-                "    {} {}",
-                style("└─").dim(),
-                style(format!("... and {remaining} more")).dim()
-            );
-        }
-    }
-    println!();
-
-    Ok(())
+    source_ops::validate_and_print(app, &result.name, Some(source_ops::MAX_TABLES_PER_SCHEMA)).await
 }
 
 async fn run_next_steps(
