+++ docs/getting-started/quickstart.mdx
@@ -28,10 +28,7 @@ coral source add github
 
 Coral prompts for any required variables or secrets. Once connected, the source's data is available as SQL tables. To update a source's credentials later, run the same command again.
 
-<Tip>
-  Run `coral source test github` to validate the connection and see the tables
-  the source exposes.
-</Tip>
+<Tip>You can re-validate anytime with `coral source test github`.</Tip>
 
 ## 3. Query your data
