+++ docs/reference/cli-reference.mdx
@@ -46,6 +46,7 @@ coral source add github
 ```
 
 Coral prompts for required variables or secrets interactively. If the source is already installed, running this command again replaces its stored credentials with the new values you provide.
+After installation, it automatically validates the connection and prints the discovered tables.
 
 ### `coral source add --file <PATH>`
